requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.4
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.8
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
frozenlist==1.8.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
multidict==6.7.0
packaging==25.0
pillow==12.0.0
pip==23.0.1
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyodbc==5.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
setuptools==65.5.0
six==1.17.0
sqlparse==0.5.3
twilio==9.8.7
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
